refactor : Controller 및 Domain 책임 분리
